Enable clang-format in Dev11_0437519_container_requirements and VSO_0191296_allocator_construct WITH MACRO CHANGES, results are acceptable.
In GENERATE_CONTAINER_TYPEDEFS, reorder opt_typename so that the usage automatically lines up nicely. This is also the order in which the macro parameters are used.

In DEFINE_TEST_IMPL, add a semicolon after preface_text to avoid confusing clang-format. This results in an unnecessary semicolon being emitted, which is fine.

In DEFINE_TEST and DEFINE_TEST_SPECIALIZATION, drop the verbose comments naming each macro parameter. DEFINE_TEST_IMPL is immediately above, so these comments provided little value.

In DEFINE_TEST_SPECIALIZATION, rename specialized_tag to my_tag. This avoids wrapping, and is still clear.