Deduction failures in pointer interconvertibilty trait tests
An upcoming MSVC version will fix a bug that caused us to accept some
template argument deduction cases that should be deduction failures.
Specifically, with a template parameter `R (C::*)` where `C` was
explicity specified you can't deduce `R` from a pointer-to-member with a
mismatched class type, even one that has an implicit conversion.

The pointer interconvertibility trait tests included a few of these
situations where a template argument was explicitly specified but
a pointer-to-member of a base class type was provided.