Remove dead code from VSO_0121275_filesystem_canonical_should_handle_many_double_dots.
This Experimental-only test was defining exec_test_output_path_too_long_behavior() but never calling it. As this involves creating symlinks,
it couldn't run for PR/CI checks anyways. We have some Standard test coverage for symlinks that warns (without failing) when
symlinks can't be created, making it possible for manual runs to laboriously exercise that code. That isn't worth the effort for the
Experimental implementation, which we aren't messing with anymore.

This function was the only user of MAX_PATH and `<Windows.h>`, which we can now drop from the test too.

Also, cleanup an outdated test comment. VSO-158882 "<filesystem>: canonical() should handle symlinks and other OS behavior" was fixed by MSVC-PR-107664 on 2018-02-23,
making the Standard implementation call GetFinalPathNameByHandleW. But this is a test for the Experimental implementation, which never calls GetFinalPathNameByHandleW, so the comment should be cauterized.