Update src/filesys.cpp to compile standalone.
Add a TRANSITION, ABI comment.

We don't need to define the silencing macro, because we can't include `<experimental/filesystem>` anymore.

We use `wstring`, so we need to include `<string>`. (This was previously dragged in, at least partially.)

We use `uint64_t` and `uintmax_t`, so we should include `<cstdint>` to be good kitties.

We need the `_MAX_FILESYS_NAME`, `_FS_BEGIN`, and `_FS_END` macros.

Within the namespace, we need `file_type`, `perms` and its `_BITMASK_OPS`, and `space_info`.