Fix ABI, part 2: `_Measure_string_prefix_iterator_utf` => `_Measure_string_prefix_iterator_utf2`
This contains `_Grapheme_break_property_iterator2<_CharT> _WrappedIter;`.

This is the final part. The alias `_Measure_string_prefix_iterator` can expand to this type,
but aliases don't need to be renamed. The only use of that alias is within the function
`_Measure_string_prefix()`, so renaming these two classes is sufficient to preserve ABI.