Improve `<coroutine>` warnings.
Make the warning for /await sound more dire by saying that `<coroutine>` is "incompatible" instead of merely "not available".

Emphasize that it is "the deprecated /await compiler option", making it clear that it should be removed.

Don't say "or use /await:strict for standard coroutines". If the user is in C++20 or later mode, removing deprecated /await is sufficient.
If they're in C++14/17 mode, warning STL4038 immediately below will tell them about /await:strict.

Don't tell users to use `<experimental/coroutine>`.

Slightly clarify warning STL4038, explaining that /await:strict is for C++14/17.