Overhaul the old _Stodx family.
_Stodx was superseded by _Stodx_v2 in "The Iostreams Floating-Point Correctness/Performance Fix To End All Fixes, WCFB01 Edition",
https://vstfdevdiv.corp.microsoft.com/DevDiv2/DevDiv/_versionControl/changeset/1572756 , merged 2016-02-04 for VS 2015 Update 2.

This fixed a huge number of correctness issues, including round-trip failures, and made `stream >> dbl` 19.5x faster and `stream >> flt` 17.0x faster.

At the time, I was aggressive enough to release this behavioral change in an update, and was vindicated with no user complaints.
However, I was very unfamiliar with floating-point, iostreams, and making significant changes to Dinkumware's legacy code.
So I added header-only _Stodx_v2 instead of changing separately compiled _Stodx.

Today, I still don't understand iostreams very well, but I am much more familiar with floating-point, and we have a decade of ABI-compat experience,
including significant changes to Dinkumware legacy code (e.g. #5473).

The interface of the _Stodx family is simple: it primarily returns a floating-point value, reads a string, potentially updates an end pointer,
takes a power of 10, and writes an error code as an out-param. The _Stodx_v2 approach can be used as a drop-in replacement.

Currently, I expect that the amount of VS 2015 RTM/Update 1 separately compiled code is extremely low. Anything compiled with VS 2015 Update 2
or the final Update 3 will be completely unaffected by this. For bincompat with RTM/Update 1, if any
such user binaries exist, this may be a minor behavioral change for them (when running against an updated VCRedist), as in we may return
floating-point values that are different by an ULP or so. It won't totally backport
the correctness fixes (which also involved iostreams preparing larger buffers for the conversions). However, due to the simplicity of the
_Stodx interface, I am confident that we can replace its guts without breaking old user binaries. Unlike the more complicated iostreams code
that calls this, _Stodx itself doesn't have unusual preconditions or postconditions.

Note that #3364 on 2023-05-04 further replaced _Stodx_v2 with _Stodx_v3, dropping the bogusly-designed "power of 10" logic.
For this PR, I used the older _Stodx_v2 approach to match the oldest _Stodx interface. (I also kept _Stodx's use of `long` for
the power of 10.)

I further suspect that the wide _WStodx family was dead even in VS 2015 RTM (I performed even older refactorings), but haven't done
the archaeology to confirm that. Since it's very easy to just replace `strtod` with `wcstod` etc., I've chosen to do that.