Add missing _EXTERN_C_UNLESS_PURE.
This was present in xfvalues.cpp but not the others.

It didn't affect anything (i.e. these "constants" were still exported as extern "C" unless pure) because xmath.hpp declared them
within _EXTERN_C_UNLESS_PURE. However, in general we try to keep definitions matching declarations, which avoids unintentional changes
during refactoring, and makes it easier to see the nature of a thing.